Update package.json name + move Item, List, Loading and TopNavigation to sub-directory component
